[mtouch/mmp] Share a few P/Invokes.
This is just moving code, no functional changes.
